chore: remove legacy codes
