+++ packages/email-auth/__tests__/index.spec.ts
@@ -1,68 +1,132 @@
 import { Chance } from 'chance';
-
-import { setup, SetupContext, teardown } from './setup';
+import getPort from 'get-port';
+import supertest from 'supertest';
+import { MongoMemoryServer } from 'mongodb-memory-server';
+import { initServer } from './app';
+import { promisify } from 'util';
+import { Otp, User } from '../src/lib/types';
+import { Document } from 'mongoose';
+import { MailService } from '@sendgrid/mail';
+import cryptoRandomString from 'crypto-random-string';
 
 const chance = new Chance();
 
-type Context = SetupContext;
+async function setup() {
+  const mongod = await MongoMemoryServer.create();
 
-beforeEach(async function (this: Context) {
-  await setup.apply(this);
-});
+  const port = await getPort();
 
-afterEach(async function (this: Context) {
-  await teardown.apply(this);
-});
+  try {
+    const serverData = await initServer(mongod.getUri());
+    const mongoDb = serverData.mongooseConnection;
+    const server = serverData.server.listen(port);
 
-describe('EmailAuthentication', () => {
-  describe('#sendEmail', () => {
-    it('should throw an error when `to` is not supplied', async function (this: Context) {
-      const response = await this.request.post('/generateOtp').send({
-        message: {},
+    const request = supertest(`http://localhost:${port}`);
+
+    return {
+      port,
+      request,
+      server,
+      mongod,
+      mongoDb,
+    };
+  } catch (error) {
+    console.log(error);
+    throw error;
+  }
+}
+
+async function teardown(ctx: Awaited<ReturnType<typeof setup>>) {
+  if (ctx.server) {
+    if (ctx.server) {
+      await promisify(ctx.server.close).call(ctx.server);
+    }
+  }
+
+  if (ctx.mongoDb) {
+    await ctx.mongoDb.close();
+  }
+
+  if (ctx.mongod) {
+    await ctx.mongod.stop();
+  }
+}
+
+describe('EmailAuthServer', () => {
+  describe('POST /otp/generate', () => {
+    it('should generate otp', async function () {
+      const ctx = await setup();
+
+      const userModel = ctx.mongoDb.model<Document & User>('User');
+
+      const emailAddress = chance.email();
+
+      await userModel.create({
+        _id: Buffer.from(chance.apple_token()),
+        emailAddress,
       });
 
-      expect(response.body).toStrictEqual({
-        message: '`to` should be supplied',
-      });
-    });
-  });
-
-  describe('#validateOtp', () => {
-    it('should throw an error when `email` is empty', async function (this: Context) {
-      const response = await this.request
-        .post('/validateOtp')
-        .send({
-          otp: '123456',
+      const sendSpy = jest
+        .spyOn(MailService.prototype, 'send')
+        .mockImplementation(async () => {
+          return [
+            {
+              statusCode: 200,
+              body: { ok: true },
+              headers: {
+                'Content-Type': 'application/json',
+              },
+            },
+            {},
+          ];
         });
 
-      expect(response.status).toBe(400);
-      expect(response.body).toStrictEqual({
-        message: '`email` should be provided',
-      });
-    });
+      await ctx.request
+        .post('/otp/generate')
+        .send({
+          emailAddress,
+        })
+        .expect(200);
 
-    it('should throw an error when `otp is empty', async function (this: Context) {
-      const emailAdress = chance.email();
-
-      const response = await this.request
-        .post('/validateOtp')
-        .send({ email: emailAdress });
-
-      expect(response.status).toBe(400);
-      expect(response.body).toStrictEqual({
-        message: '`otp` should be provided',
-      });
+      expect(sendSpy).toBeCalledTimes(1);
+      await teardown(ctx);
     });
   });
 
-  describe('#defaultResponse', () => {
-    it('should return default response', async function (this: Context) {
-      const response = await this.request.get('/');
+  describe('POST /otp/validate', () => {
+    it('should validate otp', async function () {
+      const ctx = await setup();
 
-      expect(response.status).toBe(404);
-      expect(response.body).toStrictEqual({
-        message: 'not found',
+      const userModel = ctx.mongoDb.model<Document & User>('User');
+
+      const otpModel = ctx.mongoDb.model<Document & Otp>('Otp');
+
+      const emailAddress = chance.email();
+
+      const user = await userModel.create({
+        _id: Buffer.from(chance.apple_token()),
+        emailAddress,
       });
+
+      const otp = cryptoRandomString({
+        length: 6,
+        type: 'numeric',
+      });
+
+      await otpModel.create({
+        user: user._id,
+        otp,
+      });
+
+      const response = await ctx.request
+        .post('/otp/validate')
+        .send({
+          otp,
+        })
+        .expect(200);
+
+      expect(response.body).toHaveProperty(['ok'], true);
+      await teardown(ctx);
     });
   });
 });
