+++ packages/email-auth/src/adapters/index.ts
@@ -1,3 +1,2 @@
-export { MongooseAdapter } from './mongoose-adapter';
-export { SendGridAdapter } from './send-grid-adapter';
 export { MongooseStorageAdapter } from './mongoose-storage-adapter';
+export { SendGridEmailAdapter } from './send-grid-email-adapter';
