+++ packages/email-auth/src/index.ts
@@ -1,5 +1,4 @@
 export * from './adapters';
 export * from './interfaces';
-export * from './types';
-export { EmailAuthentication } from './lib/email-authentication';
+export * from './lib/types';
 export { EmailAuthServer } from './email-auth-server';
