--- packages/email-auth/src/interfaces/emailable-provider-adapter.ts
@@ -1,6 +1,0 @@
-export interface EmailableProviderAdapter<TMessage extends {} = {}, TResponse extends {} = {}> {
-  setApiKey(key: string): void;
-  sendEmail(message: TMessage): Promise<TResponse>;
-  senderEmail: string;
-  senderName: string;
-}
