+++ packages/email-auth/src/interfaces/index.ts
@@ -1,4 +1,2 @@
-export { EmailableProviderAdapter } from './emailable-provider-adapter';
-export { PersistenceAdapter } from './persistence-adapter';
 export { EmailAdapter } from './email-adapter';
 export { StorageAdapter } from './storage-adapter';
