--- packages/email-auth/src/types/index.ts
@@ -1,6 +1,0 @@
-export type ID = Buffer;
-
-export type SenderInfo = {
-  name: string;
-  email: string;
-};
